// src/lib/types.ts
interface Student {
  studentId: string;
  firstName: string;
  lastName: string;
  program: "CPE" | "ISNE";
  courses?: string[];
  interests?: string[];
  emails?: StudentEmail[];
}
export type { Student };

interface StudentEmail {
  address: string;
}
export type { StudentEmail };

interface Instructor {
  name: string;
  email: string;
}
export type { Instructor };

interface Course {
  courseId: string;
  courseTitle: string;
  instructors: Instructor[];
  program: "CPE" | "ISNE";
  semester: "1" | "2" | "3";
  description: string;
  notifyByEmail: boolean;
}
export type { Course };

interface Enrollment {
  studentId: string;
  courseId: string;
  enrolledAt?: string;
}
export type { Enrollment };

interface User {
  username: string;
  password: string;
  studentId?: string | null;
  role: "STUDENT" | "ADMIN";
  tokens?: string[];
}
export type { User };